# Prj3/docs/openapi.yaml
openapi: 3.0.1
info:
  title: CCCD Attendance API
  version: "1.0.0"
  description: |
    REST API for device events, face comparison, employee management and admin flows.
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /api/v1/face/compare:
    post:
      summary: Server-side face compare (fallback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaceCompareRequest'
      responses:
        '200':
          description: Compare result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceCompareResponse'
...
components:
  schemas:
...
    FaceCompareRequest:
      type: object
      properties:
        cccd:
          type: string
        selfieBase64:
          type: string
        selfieVector:
          type: array
          items:
            type: number
    FaceCompareResponse:
      type: object
      properties:
        result:
          type: string
          enum: [accepted, rejected, unknown]
        bestMatch:
          type: object
          properties:
            cccd:
              type: string
            score:
              type: number
        fallbackUsed:
          type: boolean
...